Separate twitter app for OAuth
We can't have one twitter dev app for posting tweets
and oauth, because there is no separation between permission
levels.

To post tweets we need write access which forces us to ask
write access to user accounts (which we don't need).

With this approach we have one app with read access to user's
accounts and another app with write access for posting tweets.